requirements_cpu.txt
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
transformers==4.33.0
huggingface-hub
flask==2.3.3
pandas==2.1.1
Pillow==10.0.1
numpy==1.23.5
scipy==1.10.1
einops==0.6.1
cmake==3.27.4.1
timm==0.9.7
albumentations==1.3.1
imgaug==0.4.0
fvcore==0.1.5.post20221221
